Update view page style
I'm actually a bit partial on the color of the progress-bar on the view
page. It almost doesn't fit the aesthetics for the site - it sticks out
too much. Normally, with my mellow(ish) site design, colours are quite
muted, and buttons are the most bright things; even the buttons aren't
too bright, but a bit dim, allowing for a high-contrast design, that
is easier for the eyes.

Since the progress bar is very bright, it may be better to dampen it.
Or don't. That works too. I need a better opinion.