Fix sterilization issue (#12)
Pathnames from URL weren't sterilized before, leading to possible
issues such as command injection. With this, since the session ID is
predetermined, it is now much harder to do command injection. It
ain't perfect.

We use a list of registered session ID to confirm whether the sent
session ID is valid or not. The sesion IDs are added and removed
whenever a new session is started; all session IDs are generated
by python `tempfile.mkdtemp`; all session IDs are deleted by checking
that said session ID exists on the list. Thus, this makes it extremely
difficult to perform command line injection, unless one somehow modifies
the underlying code of a python core library maliciously, or the
RAM of a hosting computer. And if that happens, this is the least of
your worries.